Fix compiling templates under Perl 5.26 without '.' in @INC
Perl 5.26 changes the default contents of @INC to no longer include
".", which is required for "require q[foo.tpl]" semantics.

This breakage is invisible in tests  due to a workaround P5P put in
place to avoid breaking tests, by setting PERL_USE_UNSAFE_INC=1 in
Test::Harness and CPAN installers, which makes '.' re-appear in @INC
during installation, while still being silently broken in production.

This fix works by relying on the mechanic of require() to *NOT* consult
@INC if $path is either an absolute, or a "./" prefixed relative path,
by converting it to an absolute path just before requiring it.

Bug: https://rt.cpan.org/Ticket/Display.html?id=123380
Bug: https://bugs.gentoo.org/635464